issue #24 leads to failing regression tests - fix for test
issue #24 added the `units` keyword to latex Writers.
As an example, two columns were added to latexdicts['template'].
This template is used in the regression tests, but without matching column
names. Thus, a line of empty headers is printed. This commit adds these
empty header line to the expected output.