use fixed width information in DAOPhot tables
DAOPhot uses a fixed width format. In some cases columns touch each other
with out a space in between. In this case ``asciitable`` cannot split
them correctly. This patch makes full use of format information in the
file header.
To do so, some (minor) changes in Continuation lines inputter are required.
Before, this did some white space stripping, but really, that should not happen
in process_lines, because it makes the fixed width impossible.